requirements.txt
paddlepaddle<=2.3.0
scipy>=1.8.1
protobuf<=3.20.1
networkx>=2.5
qcompute
matplotlib
tqdm
openfermion
pyscf; platform_system == "Linux" or platform_system == "Darwin"
scikit-learn
opencv-python
